Update to run unit tests via selenium.
- Update to move the unit tests to their own function test file.
- Update to provide for allowing port specification so that tests can be run
  on specified ports to avoid collisions and make for predictable urls.
- Fix failing IE tests
- Fix the setting up of the json loading paths to work in IE since it didn't
  support location.origin.
- Provide a new makefile target for the unit tests to be started, run, and
  marked as failing/not.
- Update the ci-check to run the unit tests.
- Begin to update the browser-testing docs for the way we're currently looking
  to run them with example commands.
