+++ docs/browser-testing.rst
@@ -12,21 +12,77 @@ GUI charm and then running tests against it.
 Setting up
 ==========
 
-Environment
------------
-
-In order to run the tests you must have a Juju environment named
-"juju-gui-testing".  Consult the Juju documentation for how to set up
-an environment.
-
 Building
 --------
 
-The test script has a few system dependencies::
+The testing has the same system dependencies as development. Make sure to
+follow the HACKING doc for instructions on setting up your environment.
 
-    sudo apt-get install python-shelltoolbox python-selenium python-yaml juju
 
-It does not require that the Makefile be run.
+Manually running the browser tests
+===================================
+
+Note that there are four valid browser targets currently.
+
+  - chrome
+  - ie
+  - firefox
+  - local-firefox
+
+These are set in the below commands via an ENV variable
+`JUJU_GUI_TEST_BROWSER`. The `PORT` ENV variable is also important to help
+start the web server that serves out the test files.
+
+The commands below also mention a `[whatismyip]`. In order for SauceLabs to
+run the tests the server needs to be exposed externally via the port
+specified. In one use case, you can port forward from your router to your
+local computer and then update the `[whatismyip]` to be the actual external
+address of your home network.
+
+You can also setup things through external server with handy tactics such as
+ssh tunnels, but direct exposure is noted below as the simplest use case.
+
+
+Running Functional Tests
+------------------------
+
+There are a series of functional tests that use Selenium in the Gui. They are
+run via the Python API to Selenium and may be run against any local browser as
+well as remote browsers provided by SauceLabs.
+
+Below is a sample command to run the functional tests against the SauceLabs
+Firefox browser. If you wanted to run it against a local Firefox instance, you
+could change the `JUJU_GUI_TEST_BROWSER` to be `local-firefox`.
+
+::
+
+    APP_URL=http://[whatismyip]:8889 PORT=8889 JUJU_GUI_TEST_BROWSER="firefox" make test-browser
+
+
+Running the unit test suite a browser
+-------------------------------------
+
+You can run the prod tests in any browser locally by running the test server.
+
+::
+
+    PORT=8889 node ./test-server.js prod $(PORT)
+
+At this point you will get a url that you can point your browser to and it'll
+run against the `build-prod` version of the Gui source.
+
+
+You can also run these in SauceLabs manually by using the following:
+
+::
+
+    APP_URL=http://[whatismyip]:8889 PORT=8889 JUJU_GUI_TEST_BROWSER="chrome" make test-browser-mocha
+
+
+
+===================================================================
+BitRotted Instructions that require updating into the new CI world.
+===================================================================
 
 
 Running the tests
