+++ test-server.js
@@ -46,5 +46,6 @@ app.get('/juju-ui/:file', function(req, res) {
 
 var server = http.createServer(app);
 // Bind 0 requests a random port which we read back.
-server.listen(0);
+var port = process.env.PORT || 0;
+server.listen(process.env.PORT);
 console.log('http://0.0.0.0:' + server.address().port + '/test/index.html');
